Restructure package hierarchy  - Moves drivers to dedicated package  - Moves ASCII art to dedicated resources package
