+++ kernel/src/drivers/mod.rs
@@ -1,0 +1,5 @@
+#[macro_use]
+pub mod vga;
+
+pub mod ps2;
+pub mod ps2_io;
