+++ kernel/src/lib.rs
@@ -16,46 +16,44 @@ extern crate spin;
 mod lang;
 mod io;
 #[macro_use]
-mod vga;
-mod ps2_io;
-mod ps2;
+mod drivers;
 
-use vga::{VgaColor, Color};
+use drivers::vga::{VgaColor, Color};
 
-const FLOWER: &'static str = include_str!("flower.txt");
-const FLOWER_STEM: &'static str = include_str!("flower_stem.txt");
+const FLOWER: &'static str = include_str!("resources/ascii_art/flower.txt");
+const FLOWER_STEM: &'static str = include_str!("resources/ascii_art/flower_stem.txt");
 
 /// Kernel main function
 #[no_mangle]
 pub extern fn kmain() -> ! {
-    vga::WRITER.lock().fill_screen(Color::Black);
+    drivers::vga::WRITER.lock().fill_screen(Color::Black);
 
     // Print flower
-    vga::WRITER.lock().set_color(
+    drivers::vga::WRITER.lock().set_color(
         VgaColor::new(Color::LightBlue, Color::Black)
     );
     print!("\n{}", FLOWER);
-    vga::WRITER.lock().set_color(
+    drivers::vga::WRITER.lock().set_color(
         VgaColor::new(Color::Green, Color::Black)
     );
     print!("{}", FLOWER_STEM);
 
     // Reset colors
-    vga::WRITER.lock().set_color(
+    drivers::vga::WRITER.lock().set_color(
         VgaColor::new(Color::White, Color::Black)
     );
 
     // Reset cursor position to (0, 0)
     // It's hackish but it looks better
-    vga::WRITER.lock().set_cursor_pos((0, 0));
+    drivers::vga::WRITER.lock().set_cursor_pos((0, 0));
 
     // Print boot message
-    vga::WRITER.lock().write_str_colored(
+    drivers::vga::WRITER.lock().write_str_colored(
         "Flower kernel boot!\n-------------------\n\n",
          VgaColor::new(Color::Green, Color::Black)
     ).expect("Color code should be valid");
 
-    ps2::PS2.lock().initialize();
+    drivers::ps2::PS2.lock().initialize();
 
     halt()
 }
