+++ kernel/src/resources/ascii_art/flower.txt
