+++ kernel/src/resources/ascii_art/flower_stem.txt
